chore: cleanup and improve styling
