+++ src/pages/background/lib/handleDuplicate.ts
@@ -25,13 +25,13 @@ export default async function handleDuplicate(scheduleName: string): Promise<str
     // Extract number from parentheses and increment
     let index = match && match[2] ? parseInt(match[2].slice(1, -1), 10) + 1 : 1;
 
-    let newName = `${baseName}(${index})`;
+    let newName: string;
 
     // Increment until an unused index is found
-    // eslint-disable-next-line @typescript-eslint/no-loop-func
-    while (schedules.find(schedule => schedule.name === newName)) {
-        newName = `${baseName}(${++index})`;
-    }
+    do {
+        newName = `${baseName} (${index++})`;
+        // eslint-disable-next-line @typescript-eslint/no-loop-func
+    } while (schedules.find(schedule => schedule.name === newName));
 
     return newName;
 }
